updating for entity refactor
