+++ submitqueue/extension/changeprovider/phabricator/BUILD.bazel
@@ -12,8 +12,8 @@ go_library(
     visibility = ["//visibility:public"],
     deps = [
         "//core/metrics",
+        "//entity/change/phabricator",
         "//submitqueue/entity",
-        "//submitqueue/entity/phabricator",
         "//submitqueue/extension/changeprovider",
         "@com_github_uber_go_tally//:tally",
         "@org_uber_go_zap//:zap",
@@ -30,8 +30,9 @@ go_test(
     ],
     embed = [":phabricator"],
     deps = [
+        "//entity/change",
+        "//entity/change/phabricator",
         "//submitqueue/entity",
-        "//submitqueue/entity/phabricator",
         "//submitqueue/extension/changeprovider",
         "@com_github_stretchr_testify//assert",
         "@com_github_stretchr_testify//require",
